Organize/Modularize dashboard & file structure
