+++ client/src/App.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import MainPage from "./components/mainPage/mainPage";
-import store from "./store/store";
+import Dashboard from "./components/Dashboard/Dashboard";
 
 const App = () => {
   const state = useSelector((state) => state);
-  return <MainPage />;
+  return <Dashboard />;
 };
 
 export default App;
