+++ client/src/components/Dashboard/Orders/Orders.tsx
@@ -6,7 +6,7 @@ import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-import title from "../Title/title";
+import title from "../Title/Title";
 
 // Generate Order Data
 const createData = (
