+++ client/src/components/General/Copyright.tsx
@@ -1,0 +1,17 @@
+import { Link, Typography } from "@material-ui/core";
+import React from "react";
+
+const Copyright = () => {
+  return (
+    <Typography variant="body2" color="textSecondary" align="center">
+      {"Copyright © "}
+      <Link color="inherit" href="https://material-ui.com/">
+        Lukeylad Media{" "}
+      </Link>{" "}
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+};
+
+export default Copyright;
