+++ client/src/components/General/CustomIcons.tsx
@@ -1,0 +1,19 @@
+import { Icon, Link } from "@material-ui/core";
+import YoutubeLogo from "./YoutubeLogo.svg";
+import TwitchLogo from "./TwitchLogo.svg";
+
+export const YoutubeIcon = () => (
+  <Link href="https://www.youtube.com/">
+    <Icon>
+      <img src={YoutubeLogo} height={25} width={24} alt="description" />
+    </Icon>
+  </Link>
+);
+
+export const TwitchIcon = () => (
+  <Link href="https://www.twitch.tv/">
+    <Icon>
+      <img src={TwitchLogo} height={25} width={24} alt="description" />
+    </Icon>
+  </Link>
+);
