+++ client/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux";
-import store from "./store/store";
+import store from "./redux-store/Store";
 
 ReactDOM.render(
   <React.StrictMode>
