--- package.json
@@ -1,10 +1,0 @@
-{
-  "dependencies": {
-    "@types/redux-devtools": "^3.0.47",
-    "@types/redux-thunk": "^2.1.0",
-    "axios": "^0.21.1",
-    "react-devtools": "^4.11.0",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0"
-  }
-}
